issue: RTO after fast retransmission
In fast retransmission, we move a single segment from the unacked list to the unsent list.
But if we have RTO right after, we move all the unacked list to the unsent list, but we don't
take into consideration the we might have a retransmitted segment already in the unsent list.

Example:
unacked: 1->2->3
unsent: 4->5->6

After fast retransmission (segments 1 and 2) reorder in LWIP:
unacked: 3
unsent: 1->2->4->5->6

After RTO:
unacked: ----
unsent: 3->1->2->4->5->6

Signed-off-by: Bashar Abdelgafer <babdelgafer@nvidia.com>